+++ server/src/main/java/org/opensearch/search/aggregations/bucket/nested/ReverseNestedAggregator.java
@@ -31,7 +31,6 @@
 
 package org.opensearch.search.aggregations.bucket.nested;
 
-import com.carrotsearch.hppc.LongIntHashMap;
 import org.apache.lucene.index.LeafReaderContext;
 import org.apache.lucene.search.DocIdSetIterator;
 import org.apache.lucene.search.Query;
@@ -51,6 +50,7 @@ import org.opensearch.search.aggregations.bucket.SingleBucketAggregator;
 import org.opensearch.search.internal.SearchContext;
 
 import java.io.IOException;
+import java.util.HashMap;
 import java.util.Map;
 
 /**
@@ -91,7 +91,7 @@ public class ReverseNestedAggregator extends BucketsAggregator implements Single
         if (parentDocs == null) {
             return LeafBucketCollector.NO_OP_COLLECTOR;
         }
-        final LongIntHashMap bucketOrdToLastCollectedParentDoc = new LongIntHashMap(32);
+        final Map<Long, Integer> bucketOrdToLastCollectedParentDoc = new HashMap<>(32);
         return new LeafBucketCollectorBase(sub, null) {
             @Override
             public void collect(int childDoc, long bucket) throws IOException {
@@ -99,16 +99,15 @@ public class ReverseNestedAggregator extends BucketsAggregator implements Single
                 final int parentDoc = parentDocs.nextSetBit(childDoc);
                 assert childDoc <= parentDoc && parentDoc != DocIdSetIterator.NO_MORE_DOCS;
 
-                int keySlot = bucketOrdToLastCollectedParentDoc.indexOf(bucket);
-                if (bucketOrdToLastCollectedParentDoc.indexExists(keySlot)) {
-                    int lastCollectedParentDoc = bucketOrdToLastCollectedParentDoc.indexGet(keySlot);
+                Integer lastCollectedParentDoc = bucketOrdToLastCollectedParentDoc.get(bucket);
+                if (lastCollectedParentDoc != null) {
                     if (parentDoc > lastCollectedParentDoc) {
                         collectBucket(sub, parentDoc, bucket);
-                        bucketOrdToLastCollectedParentDoc.indexReplace(keySlot, parentDoc);
+                        bucketOrdToLastCollectedParentDoc.put(bucket, parentDoc);
                     }
                 } else {
                     collectBucket(sub, parentDoc, bucket);
-                    bucketOrdToLastCollectedParentDoc.indexInsert(keySlot, bucket, parentDoc);
+                    bucketOrdToLastCollectedParentDoc.put(bucket, parentDoc);
                 }
             }
         };