+++ server/src/test/java/org/opensearch/search/aggregations/bucket/range/BinaryRangeAggregatorTests.java
@@ -45,8 +45,6 @@ import org.opensearch.search.aggregations.bucket.range.BinaryRangeAggregator.Sor
 import org.opensearch.search.aggregations.bucket.range.BinaryRangeAggregator.SortedSetRangeLeafCollector;
 import org.opensearch.test.OpenSearchTestCase;
 
-import com.carrotsearch.hppc.LongHashSet;
-
 public class BinaryRangeAggregatorTests extends OpenSearchTestCase {
 
     private static class FakeSortedSetDocValues extends AbstractSortedSetDocValues {
@@ -121,12 +119,12 @@ public class BinaryRangeAggregatorTests extends OpenSearchTestCase {
         final int[] expectedCounts = new int[ranges.length];
         final int maxDoc = randomIntBetween(5, 10);
         for (int doc = 0; doc < maxDoc; ++doc) {
-            LongHashSet ordinalSet = new LongHashSet();
+            Set<Long> ordinalSet = new HashSet<>();
             final int numValues = randomInt(maxNumValuesPerDoc);
             while (ordinalSet.size() < numValues) {
-                ordinalSet.add(random().nextInt(terms.length));
+                ordinalSet.add(TestUtil.nextLong(random(), 0, terms.length - 1));
             }
-            final long[] ords = ordinalSet.toArray();
+            final long[] ords = ordinalSet.stream().mapToLong(Long::longValue).toArray();
             Arrays.sort(ords);
             values.ords = ords;
 
@@ -222,12 +220,12 @@ public class BinaryRangeAggregatorTests extends OpenSearchTestCase {
         final int[] expectedCounts = new int[ranges.length];
         final int maxDoc = randomIntBetween(5, 10);
         for (int doc = 0; doc < maxDoc; ++doc) {
-            LongHashSet ordinalSet = new LongHashSet();
+            Set<Long> ordinalSet = new HashSet<>();
             final int numValues = randomInt(maxNumValuesPerDoc);
             while (ordinalSet.size() < numValues) {
-                ordinalSet.add(random().nextInt(terms.length));
+                ordinalSet.add(TestUtil.nextLong(random(), 0, terms.length - 1));
             }
-            final long[] ords = ordinalSet.toArray();
+            final long[] ords = ordinalSet.stream().mapToLong(Long::longValue).toArray();
             Arrays.sort(ords);
             values.ords = ords;